% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OfficialJoke.R
\name{get_joke}
\alias{get_joke}
\title{Retrieves jokes from API}
\usage{
get_joke(
  type = c("general", "knock-knock", "programming"),
  choice = c("random", "ten"),
  return_type = c("list", "dataframe")
)
}
\arguments{
\item{type}{The endpoint type, one of general, knock-knock or programming.
Default to general if nothing is provided.}

\item{choice}{The choice type, one of random or ten.
Default to random if nothing is provided.}

\item{return_type}{The return_type, one of list or dataframe.
Default to list if nothing is provided.}
}
\value{
List with all parameters of that joke
}
\description{
Retrieves jokes from API
}
\examples{
\dontrun{

# Retrieve random general joke,return list
joke1 <- get_joke(type = "general",choice="random",return_type="list")

# Retrieve random programming joke,return dataframe
joke2 <- get_joke(type = "programming",choice="random",return_type="dataframe")

# Retrieve ten knock-knock jokes,return list
joke3 <- get_joke(type = "knock-knock",choice="ten",return_type="list")

# Retrieve ten general jokes,return dataframe
joke4 <- get_joke(type = "general",choice="ten",return_type="dataframe")

# Retrieve data with default value
joke5 <- get_joke()
}
}
\author{
Ximing Zhang
}
