% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OfficialJoke.R
\name{get_random_joke}
\alias{get_random_joke}
\title{Retrieves random official joke from API
The Final Project for QMSS MDS 2019 Fall
Option B: API project}
\usage{
get_random_joke()
}
\value{
List with all parameters of the joke
}
\description{
Retrieves random official joke from API
The Final Project for QMSS MDS 2019 Fall
Option B: API project
}
\examples{
\dontrun{

# Retrieve random joke
rand_joke1 <- get_random_joke()

}
}
\author{
Ximing Zhang
}
