% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OfficialJoke-package.R
\docType{package}
\name{OfficialJoke-package}
\alias{OfficialJoke-package}
\alias{OfficialJokeR}
\title{R wrapper for An API of Official Joke}
\value{
List with all parameters of the joke from official joke APIs.
}
\description{
This package provides access to the \href{https://github.com/15Dkatz/official_joke_api}{official_joke_api}
API from R. Final Project For MDS 2019 Fall.
}
\examples{
get_joke()
get_random_joke()
get_joke(type = "general",choice="ten",return_type="dataframe")

}
\author{
Ximing Zhang
}
